converted ncep_qc_threshold from an integer to a real(r8)
git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@2400 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 36a9e20bfca63f0194a940a3eb8ea645b5148255

+++ filter/filter.f90
@@ -86,7 +86,7 @@ integer  :: num_output_obs_members   = 0
 integer  :: output_interval = 1
 integer  :: num_groups = 1
 real(r8) :: outlier_threshold = -1.0_r8
-integer  :: ncep_qc_threshold = 4
+real(r8) :: ncep_qc_threshold = 4.0_r8
 
 character(len = 129) :: obs_sequence_in_name  = "obs_seq.out",    &
                         obs_sequence_out_name = "obs_seq.final",  &
@@ -970,7 +970,7 @@ do j = 1, obs_ens_handle%my_num_vars
         
       ! PAR: THIS SHOULD BE IN QC MODULE 
       ! Check on the outlier threshold quality control: move to QC module
-      if(do_outlier .and. nint(obs_ens_handle%copies(OBS_GLOBAL_QC_COPY, j)) < ncep_qc_threshold) then
+      if(do_outlier .and. (obs_ens_handle%copies(OBS_GLOBAL_QC_COPY, j) < ncep_qc_threshold)) then
          obs_prior_mean = obs_ens_handle%copies(OBS_PRIOR_MEAN_START, j)
          obs_prior_var = obs_ens_handle%copies(OBS_PRIOR_VAR_START, j)
          obs_val = obs_ens_handle%copies(OBS_VAL_COPY, j)
